==> src/0-raw-data-statistics.ipynb <==
# Descriptive analysis of input data
1. Number of stops, users, active days, and time span.
2. Users' active days description: count, # of stops per active day, duration of activities per active day.

%load_ext autoreload
%autoreload 2
%cd D:\mad4abm

# Load libs
import pandas as pd
import geopandas as gpd
import sqlalchemy
import numpy as np
from scipy import stats
from tqdm import tqdm
from lib import preprocess as preprocess

# Data location
user = preprocess.keys_manager['database']['user']
password = preprocess.keys_manager['database']['password']
port = preprocess.keys_manager['database']['port']
db_name = preprocess.keys_manager['database']['name']
engine = sqlalchemy.create_engine(f'postgresql://{user}:{password}@localhost:{port}/{db_name}')

## 1. Load data

df = pd.read_sql_query(sql="""SELECT * FROM stops;""", con=engine)
df.head(3)

print("Data cover %s devices of %s stop points."%(df['uid'].nunique(), len(df)))

df.loc[:, 'date'] = df.loc[:, 'TimeLocal'].dt.date
print("Data have %s active days from %s to %s."%(df.loc[:, 'date'].nunique(), df.loc[:, 'date'].min(), df.loc[:, 'date'].max()))

df.loc[:, 'dur'] = df['leaving_datetime'] - df['datetime']
tqdm.pandas()
df.loc[:, 'dur'] = df.loc[:, 'dur'].progress_apply(lambda x: x.total_seconds())

### 1.1 Overall statistics by month

tqdm.pandas()
df.loc[:, 'month'] = df.loc[:, 'date'].progress_apply(lambda x: str(x).split('-')[1])
df.groupby('month').progress_apply(lambda data: pd.Series({'num_stops': len(data), 'num_uid': data['uid'].nunique()}))

## 2. User statistics

df_test = df.loc[df.uid == df.uid[0], :]
df_test.head()

def user_stats(data):
    # No. of active days
    num_days = data.loc[:, 'date'].nunique()
    # No. of stays
    num_stays = len(data)
    # Description of active days
    ## Median value of stays/active day
    num_stays_act = data['date'].value_counts().median()
    ## Median value of total duration of stays/active day (h)
    dur_total_act = np.median(data.groupby('date')['dur'].sum()/60/60)
    ## Median value of duration/stay (min)
    dur_median = data['dur'].median()/60
    return pd.Series(dict(num_days=num_days, num_stays=num_stays, num_stays_act=num_stays_act,
                          dur_total_act=dur_total_act, dur_median=dur_median))
tqdm.pandas()
df_user = df.groupby('uid').progress_apply(user_stats).reset_index()

df_user.describe()

preprocess.dump2db_df(df_user, user, password, port, db_name, table_name='stops', schema_name='description')

## 3. Home locations

df_home = pd.read_sql_query(sql="""SELECT * FROM vgr_homes_2019;""", con=engine)
df_home.head(3)

print("Homes cover %s devices."%len(df_home))

df_home_desc = df_home['deso_code'].value_counts().rename(index='count').reset_index().rename(columns={'index': 'deso'})

### 3.1 Population representativeness

gdf = gpd.GeoDataFrame.from_postgis(sql="""SELECT * FROM zones;""", con=engine)
gdf = gdf.loc[:, ['deso', 'befolkning', 'geom']]
gdf_home_desc = gdf.merge(df_home_desc, on='deso')

gdf_home_desc.to_file('results/zones_vgr_homes_2019.shp')

stats.kendalltau(gdf_home_desc['befolkning'], gdf_home_desc['count'])
