==> src/0-survey-trips-processing.ipynb <==
# Covert trips into activities and create commuting matrices
1. Check typical trip chains of the travel survey.
2. Design filters for preparing the data for activities.
3. Extract activities from trip chains.

%load_ext autoreload
%autoreload 2
%cd D:\mad4abm

import pandas as pd
import numpy as np
from datetime import datetime
import datetime as dt
from tqdm import tqdm

df = pd.read_csv("dbs/survey/day_trips.csv", encoding="latin-1")
df = df.drop_duplicates(subset=['sub_id', 'trip_id', 'trip_main_id'])
df = df.loc[:, ["sub_id", "date", 'trip_id', "trip_main_id", 'purpose_main', 'purpose', 'origin_main_time', 'desti_main_time', 'origin_main_deso',  'desti_main_deso', 'trip_main_mode']]
df = df.sort_values(by=['sub_id', 'trip_id', 'trip_main_id'])
print(df.sub_id.nunique())

Drop those participants with incomplete information.

def completeness_check(data):
    if data.isnull().values.any():
        return pd.Series({'complt': 0})
    return pd.Series({'complt': 1})

tqdm.pandas()
df_complt = df.groupby("sub_id").progress_apply(completeness_check).reset_index()

df = df.loc[df.sub_id.isin(df_complt.loc[df_complt.complt == 1, 'sub_id']), :]
for var in ['purpose', 'purpose_main', 'origin_main_time', 'desti_main_time', 'trip_main_mode']:
    df.loc[:, var] = df.loc[:, var].astype(int)
print('Valid participants: %s'%df.sub_id.nunique())
df.head()

purpose_main 2, 3, the rest = commuting (XW), School (S), and Other (O)

Recoding it to 2, 3, and 0

df.loc[:, 'purpose_main'] = df.loc[:, 'purpose_main'].apply(lambda x: 0 if x not in (2, 3) else x)

## 1. Trip chains

df_test = df.loc[df['sub_id'] == 20110111013, :]
df_test

df_test = df.loc[df['sub_id'] == 20110111036, :]
df_test

df_test = df.loc[df['sub_id'] == 20110111049, :]
df_test

df_test = df.loc[df['sub_id'] == 20110111072, :]
df_test

df_test = df.loc[df['sub_id'] == 20110121017, :]
df_test

## 2. Extract individuals for activities

The individuals satisfying a combination of the below conditions will be selected for the activity extraction.

1. The last trip is commuting.
2. The last trip is a round trip.
3. The last trip and the same trip has the same purpose.
4. The first origin and the last destination are in the same DeSO zone

def individual_process(data):
    first_purpose = data['purpose'].values[0]
    last_purpose = data['purpose'].values[-1]
    last_trip_num = data['trip_id'].max()
    last_round_trip = int(len(data.loc[data['trip_id'] == last_trip_num, :]) > 1)
    same_zone = int(data['origin_main_deso'].values[0] == data['desti_main_deso'].values[-1])
    comm_end = int(last_purpose == 2)
    round_end = int(last_round_trip == 1)
    same_pur = int(first_purpose == last_purpose)
    return pd.Series({'same_zone':same_zone, 'comm_end':comm_end, 'round_end':round_end, 'same_pur': same_pur})

df_act_patterns = df.groupby('sub_id').apply(individual_process).reset_index()
df_act_patterns.head()

df_act_patterns.loc[:, 'any'] = df_act_patterns.apply(lambda row: int(row['comm_end'] + row['round_end'] + row['same_pur'] > 0), axis=1)

share_same_zone = df_act_patterns.loc[df_act_patterns.same_zone == 1, :].size / df_act_patterns.size *100
share_comm_end = df_act_patterns.loc[df_act_patterns.comm_end == 1, :].size / df_act_patterns.size *100
share_round_end = df_act_patterns.loc[df_act_patterns.round_end == 1, :].size / df_act_patterns.size *100
share_same_pur = df_act_patterns.loc[df_act_patterns.same_pur == 1, :].size / df_act_patterns.size *100
print("Share of participants with same start and end DeSO zone: %.2f"%share_same_zone)
print("Share of participants with same start and end purpose: %.2f"%share_same_pur)
print("Share of participants with last trip being commuting: %.2f"%share_comm_end)
print("Share of participants with last trip being a round trip: %.2f"%share_round_end)

### 2.1 Participants filter
Participants with same start and end DeSO zone AND any of the three conditions is met: same start and end purpose, last trip being commuting, and last trip being a round trip.

selected_inds = df_act_patterns.loc[(df_act_patterns.same_zone == 1) & (df_act_patterns['any'] == 1), 'sub_id'].values
print("Share of participants selected: %.2f"%(len(selected_inds) / len(df_act_patterns) * 100))

df_sub = df.loc[df.sub_id.isin(selected_inds), :].copy()

## 3. Convert trips into time series of activities
### 3.1 Time processing

df_sub.head()

def time_padding(x):
    if len(x) == 4:
        return x
    if len(x) == 3:
        return '0' + x
    if len(x) == 2:
        return '00' + x
    if len(x) == 1:
        return '000' + x

df_sub.loc[:, 'origin_main_T'] = df_sub['date'] + ' ' + df_sub['origin_main_time'].astype(str).apply(lambda x: time_padding(x))
df_sub.loc[:, 'desti_main_T'] = df_sub['date'] + ' ' + df_sub['desti_main_time'].astype(str).apply(lambda x: time_padding(x))

df_sub.loc[:, 'origin_main_T'] = df_sub['origin_main_T'].apply(lambda x: datetime.strptime(str(x),'%Y-%m-%d %H%M'))
df_sub.loc[:, 'desti_main_T'] = df_sub['desti_main_T'].apply(lambda x: datetime.strptime(str(x),'%Y-%m-%d %H%M'))

df_sub.loc[:, 'travel_time'] = df_sub.loc[:, 'desti_main_T'] - df_sub.loc[:, 'origin_main_T']
df_sub.loc[df_sub.travel_time < dt.timedelta(minutes=0), 'desti_main_T'] += dt.timedelta(minutes=1440)
df_sub.loc[:, 'travel_time'] = df_sub.loc[:, 'desti_main_T'] - df_sub.loc[:, 'origin_main_T']
df_sub.loc[:, 'travel_time'] /= dt.timedelta(minutes=1)

### 3.2 Create time series

df_sub_example = df_sub.loc[df_sub.sub_id == 20110111013, :].copy()
df_sub_example

date_s = df_sub_example.date.iloc[0]
time_s = df_sub_example.origin_main_T.iloc[0]
time_e = df_sub_example.desti_main_T.iloc[-1]
date_e = str(time_e).split(' ')[0]

date_s, time_s, time_e, date_e

if date_s == date_e:
    # The date's 00:00
    time_start = datetime.strptime(str(date_s),'%Y-%m-%d')
    # The next date's 00:00
    date_e_r = str(time_e + dt.timedelta(minutes=1440)).split(' ')[0]
    time_end = datetime.strptime(date_e_r,'%Y-%m-%d')
else:
    # 24 hours before the end date
    date_s_r = str(time_e - dt.timedelta(minutes=1440)).split(' ')[0]
    time_start = datetime.strptime(str(date_s_r),'%Y-%m-%d')
    time_end = datetime.strptime(date_e,'%Y-%m-%d')
activity_start_times = np.insert(df_sub_example.desti_main_T.values, 0, time_start, axis=0)
activity_end_times = np.insert(df_sub_example.origin_main_T.values, len(df_sub_example), time_end, axis=0)

purposes = np.insert(df_sub_example.purpose_main.values, 0, 0, axis=0)
purposes[len(purposes)-1] = 0

df_act = pd.DataFrame()
df_act.loc[:, 'act_start'] = activity_start_times
df_act.loc[:, 'act_end'] = activity_end_times
df_act.loc[:, 'purpose'] = purposes
df_act

### 3.2.1 Activity conversion

def trips2activities(data):
    # Start and end time
    date_s = data.date.iloc[0]
    time_e = data.desti_main_T.iloc[-1]
    date_e = str(time_e).split(' ')[0]
    if date_s == date_e:
        # The date's 00:00
        time_start = datetime.strptime(str(date_s),'%Y-%m-%d')
        # The next date's 00:00
        date_e_r = str(time_e + dt.timedelta(minutes=1440)).split(' ')[0]
        time_end = datetime.strptime(date_e_r,'%Y-%m-%d')
    else:
        # 24 hours before the end date
        date_s_r = str(time_e - dt.timedelta(minutes=1440)).split(' ')[0]
        time_start = datetime.strptime(str(date_s_r),'%Y-%m-%d')
        time_end = datetime.strptime(date_e,'%Y-%m-%d')
    activity_start_times = np.insert(data.desti_main_T.values, 0, time_start, axis=0)
    activity_end_times = np.insert(data.origin_main_T.values, len(data), time_end, axis=0)
    # Purpose
    purposes = np.insert(data.purpose_main.values, 0, 0, axis=0)
    purposes[len(purposes)-1] = 0
    # DeSO zone
    zones = np.insert(data.desti_main_deso.values, 0, data.origin_main_deso.iloc[0], axis=0)
    # Create a dataframe
    df_act = pd.DataFrame()
    df_act.loc[:, 'act_start'] = activity_start_times
    df_act.loc[:, 'act_end'] = activity_end_times
    df_act.loc[:, 'purpose'] = purposes
    df_act.loc[:, 'zone'] = zones
    return df_act

tqdm.pandas()
df_act = df_sub.groupby('sub_id').progress_apply(trips2activities).reset_index()

df_act = df_act.drop(columns=['level_1'])

df_act.loc[:, 'h_s'] = df_act.loc[:, 'act_start'].apply(lambda x: x.hour * 60 + x.minute)
df_act.loc[:, 'h_e'] = df_act.loc[:, 'act_end'].apply(lambda x: x.hour * 60 + x.minute)
df_act.loc[:, 'dur'] = df_act['act_end'] - df_act['act_start']
df_act.loc[:, 'dur'] = df_act.loc[:, 'dur'].apply(lambda x: x.total_seconds() / 60)
df_act.head()

# Remove those with negative activity duration
df_act = df_act.loc[~df_act.sub_id.isin(df_act.loc[df_act.dur <= 0,'sub_id'].unique()), :]

pur_dict = {0: 'Home', 2: 'Work', 3: 'School'}
df_act.loc[:, 'Purpose'] = df_act.loc[:, 'purpose'].apply(lambda x: pur_dict[x] if x in pur_dict else 'Other')

df_act.to_csv('dbs/survey/day_act.csv', index=False)

## 4. Create commuting matrices
1. Municipality-level
2. DeSO-level

df = pd.read_csv("dbs/survey/day_trips.csv", encoding="latin-1")
df = df.drop_duplicates(subset=['sub_id'])
df = df.loc[:, ["sub_id", "sub_weight", 'home_deso', 'home_municipality', 'workplace_deso', 'workplace_municipality']]
len(df)

### 4.1 Municipality level

df_mu = df[(df['home_municipality'].notna())&(df['workplace_municipality'].notna())]
len(df_mu)

df_od = df_mu.groupby(["home_municipality", "workplace_municipality"])["sub_weight"].sum().reset_index()
df_od.columns = ["ozone", "dzone", "sv_commute"]
df_od.to_csv('dbs/survey/commute_od_municipality.csv', index=False, encoding='latin-1')

### 4.2 DeSO zone level

df_mu = df[(df['home_deso'].notna())&(df['workplace_deso'].notna())]
len(df_mu)

df_od = df_mu.groupby(["home_deso", "workplace_deso"])["sub_weight"].sum().reset_index()
df_od.columns = ["ozone", "dzone", "sv_commute"]
df_od.to_csv('dbs/survey/commute_od_deso.csv', index=False, encoding='latin-1')

## 5. Create overall matrices
1. Municipality-level
2. DeSO-level

df = pd.read_csv("dbs/survey/day_trips.csv", encoding="latin-1")
df = df.drop_duplicates(subset=['sub_id'])
df = df.loc[:, ["sub_id", "sub_weight", 'origin_main_deso', 'desti_main_deso',
                'origin_main_municipality', 'desti_main_municipality']]
len(df)

### 5.1 Municipality level

df_mu = df[(df['origin_main_municipality'].notna())&(df['desti_main_municipality'].notna())]
len(df_mu)

df_od = df_mu.groupby(["origin_main_municipality", "desti_main_municipality"])["sub_weight"].sum().reset_index()
df_od.columns = ["ozone", "dzone", "sv"]
df_od = df_od.astype({'ozone':'int', 'dzone':'int'})
df_od.to_csv('dbs/survey/od_municipality.csv', index=False, encoding='latin-1')

df_mu = df[(df['origin_main_deso'].notna())&(df['desti_main_deso'].notna())]
len(df_mu)

df_od = df_mu.groupby(["origin_main_deso", "desti_main_deso"])["sub_weight"].sum().reset_index()
df_od.columns = ["ozone", "dzone", "sv"]
df_od.to_csv('dbs/survey/od_deso.csv', index=False, encoding='latin-1')
