==> src/2-home-detection.ipynb <==
# Home detection and corresponding DeSO zone

%load_ext autoreload
%autoreload 2
%cd D:\mad4abm

# Load libs
import pandas as pd
import geopandas as gpd
import sqlalchemy
import numpy as np
from scipy import stats
from tqdm import tqdm
from lib import preprocess as preprocess
from shapely.geometry import MultiPoint

# Data location
user = preprocess.keys_manager['database']['user']
password = preprocess.keys_manager['database']['password']
port = preprocess.keys_manager['database']['port']
db_name = preprocess.keys_manager['database']['name']
engine = sqlalchemy.create_engine(f'postgresql://{user}:{password}@localhost:{port}/{db_name}')

## 1. Home detection test

uid = '0d3a07aa-e73c-4f98-b10e-cb8d3da77d79'
df_eg = pd.read_sql_query(sql="""SELECT * FROM stops_subset WHERE uid = '%s';"""%uid, con=engine)
df_eg.head()

df_eg.head()

Holiday seasons: 6.23 - 8.10, and > 12.22

# Identify home location using individual stop sequence
def home_detection(data):
    def time_in_range(start, end, x):
        """Return true if x is in the range [start, end]"""
        if (x >= start) | (x < end):
            return True
        else:
            return False

    def point(row):
        start = 19
        end = 8
        h_s = row['TimeLocal'].hour
        h_e = row['leaving_TimeLocal'].hour
        pt = row['dur'] / 3600  # stay length in hour
        if row['weekday'] == 1:
            if ~time_in_range(start, end, h_s) & ~time_in_range(start, end, h_e):
                pt = 0
            if ~time_in_range(start, end, h_s) & time_in_range(start, end, h_e):
                pt = pt - (19 - h_s)
            if time_in_range(start, end, h_s) & ~time_in_range(start, end, h_e):
                pt = pt - (h_e - 8)
        return pt

    data_cat = data.reset_index(drop=True)
    data_cat.loc[:, 'point'] = data_cat.apply(lambda row: point(row), axis=1)
    try:
        if len(data_cat) > 1:
            home = pd.DataFrame(data_cat.groupby(['cluster'])['point'].sum())  # Share of stay in total
            home = home.reset_index()
            home = home.loc[home['point'].idxmax()].values
            coords = data.loc[data.cluster == home[0], ['lng', 'lat']].values
            centroid = (MultiPoint(coords).centroid.x, MultiPoint(coords).centroid.y) #x='longitude', y='latitude'
        else:
            home = [0, 0]
            centroid = (0, 0)
        return pd.Series({'cluster': home[0], 'length': home[1], 'lng': centroid[0], 'lat': centroid[1]})
    except:
        home = [0, 0, 0, 0]
        centroid = (0, 0)
        return pd.Series({'cluster': home[0], 'length': home[1], 'lng': centroid[0], 'lat': centroid[1]})

df_eg.loc[:, 'weekday'] = df_eg.loc[:, 'TimeLocal'].apply(lambda x: 1 if x.weekday() in (0, 1, 2, 3, 4) else 0)

home_detection(df_eg)

## 2. Home detection for all

df = pd.read_sql_query(sql="""SELECT * FROM stops_subset;""", con=engine)
df.loc[:, 'weekday'] = df.loc[:, 'TimeLocal'].apply(lambda x: 1 if x.weekday() in (0, 1, 2, 3, 4) else 0)

tqdm.pandas()
df_home = df.groupby('uid').progress_apply(home_detection).reset_index()
df_home.loc[:, 'cluster'] = df_home.loc[:, 'cluster'].astype(int)
df_home.head()

## 3. Get DeSO zones for detected home

gdf = gpd.GeoDataFrame.from_postgis("SELECT deso, geom FROM public.zones", con=engine).to_crs(4326)

gdf_home = preprocess.df2gdf_point(df_home, 'lng', 'lat', crs=4326, drop=True)
gdf_home = gpd.sjoin(gdf_home, gdf)
gdf_home.head()

df_home_deso = pd.merge(df_home, gdf_home.loc[:, ['uid', 'deso']], on='uid', how='inner')
df_home_deso.head()

df_home_deso.loc[df_home_deso['length']>0, :].to_sql('home', engine, schema='public', index=False, method='multi', if_exists='replace', chunksize=10000)
