==> src/4-work-home-detection.ipynb <==
# Distance-to-template home and work detection
Check the results of the distances to home and work temporal profiles from the survey

%load_ext autoreload
%autoreload 2
%cd D:\mad4abm

# Load libs
import pandas as pd
import sqlalchemy
from tqdm import tqdm
from lib import preprocess as preprocess

## 1. Load distances and get the shortest distance to work and home

# Data location
user = preprocess.keys_manager['database']['user']
password = preprocess.keys_manager['database']['password']
port = preprocess.keys_manager['database']['port']
db_name = preprocess.keys_manager['database']['name']
engine = sqlalchemy.create_engine(f'postgresql://{user}:{password}@localhost:{port}/{db_name}')

df = pd.read_sql_query(sql="""SELECT * FROM description.tempo_top3_distance;""", con=engine)
df_home = pd.read_sql_query(sql="""SELECT uid, cluster FROM home;""", con=engine)

### 1.1 Select individually the clusters with the shortest distances to home and work

def ind_select(data):
    home_ = data.loc[data.dist2home_wt == data.dist2home_wt.min(), ['cluster', 'dist2home_wt']].values[0]
    data_rest = data.loc[data.cluster != home_[0], :]
    work_ = data_rest.loc[data_rest.dist2work_wt == data_rest.dist2work_wt.min(), ['cluster', 'dist2work_wt']].values[0]
    return pd.Series(dict(home=home_[0], dist2home_wt=home_[1],
                          work=work_[0], dist2work_wt=work_[1]))
tqdm.pandas()
df_hw = df.groupby('uid').progress_apply(ind_select).reset_index().astype({"home": int, "work": int})

df_hw = pd.merge(df_hw, df_home.rename(columns={'cluster': 'home_b'}), on='uid', how='left')
df_hw.head()

len(df_hw.loc[df_hw.home == df_hw.home_b, :]) / len(df_hw) * 100

len(df_hw.loc[df_hw.work == df_hw.home_b, :]) / len(df_hw) * 100

df_hw.to_sql('home_work', engine, schema='description', index=False, method='multi', if_exists='replace', chunksize=10000)

## 2. Create data for visualisation

df_hw.loc[:, 'home_group'], _ = pd.qcut(df_hw.dist2home_wt, q=10, retbins=True)
df_hw.loc[:, 'work_group'], _ = pd.qcut(df_hw.dist2work_wt, q=10, retbins=True)
df_hw.head()

import pandas as pd
grps, bins = pd.qcut(range(0, 100), q=10, retbins=True)

df_hw_h_eg = df_hw.groupby('home_group').sample(n=1, random_state=1).loc[:, ['uid', 'home', 'dist2home_wt', 'home_group']]
df_hw_w_eg = df_hw.groupby('work_group').sample(n=1, random_state=1).loc[:, ['uid', 'work', 'dist2work_wt', 'work_group']]

agents_eg = ','.join(["'" + x + "'" for x in df_hw_h_eg['uid'].values.tolist()] +
                     ["'" + x + "'" for x in df_hw_w_eg['uid'].values.tolist()])

df_examples = pd.read_sql(f'''SELECT * FROM description.tempo_top3 WHERE uid IN ({agents_eg});''', con=engine)
df_examples.head()

### 2.1 Save the data for further visualisation

df_hw_h_eg.sort_values(by=['dist2home_wt'], ascending=False).to_csv('results/home_detection_examples.csv', index=False)
df_hw_w_eg.sort_values(by=['dist2work_wt'], ascending=False).to_csv('results/work_detection_examples.csv', index=False)
df_examples.sort_values(by=['uid', 'cluster', 'half_hour']).to_csv('results/hw_detection_temporal_examples.csv', index=False)
