==> src/6-commute-odm-comparison.ipynb <==
# ODMs comparison

%load_ext autoreload
%autoreload 2
%cd D:\mad4abm

# Load libs
import pandas as pd
import sqlalchemy
from lib import preprocess as preprocess
from lib import helpers as helpers

# Data location
user = preprocess.keys_manager['database']['user']
password = preprocess.keys_manager['database']['password']
port = preprocess.keys_manager['database']['port']
db_name = preprocess.keys_manager['database']['name']
engine = sqlalchemy.create_engine(f'postgresql://{user}:{password}@localhost:{port}/{db_name}')

df = pd.read_sql_query(sql="""SELECT * FROM odm.commute;""", con=engine)

## 1. Survey ODMs: commute vs. overall

df_muni = pd.merge(pd.read_csv('dbs/survey/od_municipality.csv'),
                   pd.read_csv('dbs/survey/commute_od_municipality.csv'),
                   on=['ozone', 'dzone'], how='left').fillna(0)
df_deso = pd.merge(pd.read_csv('dbs/survey/od_deso.csv'),
                   pd.read_csv('dbs/survey/commute_od_deso.csv'),
                   on=['ozone', 'dzone'], how='left').fillna(0)

ssi_muni = helpers.odm_ssi(df_muni, x_field='sv', y_field='sv_commute')
ssi_muni

ssi_deso = helpers.odm_ssi(df_deso, x_field='sv', y_field='sv_commute')
ssi_deso
